Add daftar pengumpulan sampah, edit daftar pengumpulan sampah activity for admin

+++ app/src/main/java/com/faridwaid/banksampahmliriprowo/Users.kt
@@ -8,7 +8,7 @@ data class Users(
     val photoProfil : String,
     val jumlahSetoran : Int,
     val jumlahPenarikan : Int,
-    val saldo : Int,
+    val saldo : Long,
     val token : String
 ){
     constructor(): this("","", "", "", 0, 0, 0, ""){
